Skip cached files during check unless hash changed
Music conversion always happens before the cache is ever built, so after the first run the cached files should already be in the new format. Then, if a filepath doesn't exist in the hash, or its hash was changed, then it can be rechecked to ensure it's still new format.