Update metadata YAML formmask format and let cache store formmask data
The original formmask format is hard to understand for newcomers, and hard to edit. This change introduces a much easier to understand and edit system, where the keys in formmask are "channel #" and "cumulative states" and their values are a YAML list of strings. This makes it easier to edit and target specific channels, with fields allowed to be left out without causing any issues.

This also makes the formmask bytearray accessible to the music cache, allowing cached files to properly use their formmask data.